requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
cloudgenix==6.3.2b1
cryptography==3.4.6
idna==2.10
ipcalc==1.99.0
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.9.0.post0
PyYAML==5.4.1
requests==2.31.0
six==1.15.0
urllib3==2.2.1
websockets==8.1
